# refactor(ext/node): Use an internal DataView in Buffer