@aapoalas 2024-11-03
I have decided that this is not a good idea: Adding something like 80 bytes to an already fairly large static size of `Buffer` AKA `Uint8Array` plus the internal `ArrayBuffer`, for what is a fairly small performance upside (and even some downsides, especially in the more necessary parts).

The proper way to improve the performance of `Buffer` read and write APIs would be to implement them in Rust.